Detect bad github key in curricula repos

+++ deploy.sh
@@ -25,7 +25,14 @@ then
     git config --global user.name "CircleCI"
     git config --global user.email "circleci@thinkful.com"
     git commit -m "automatic commit of uuids after pushing to master [CI skip]"
-    git push origin master --force
+    PUSH_MASTER_OUTPUT=$(git push origin master --force 2>&1)
+    if [[ $PUSH_MASTER_OUTPUT =~ "ERROR" ]]
+    then
+      echo $PUSH_MASTER_OUTPUT
+      echo "Aborting, error pushing to GitHub. This curriculum a read+write key in CI."
+      exit 1
+    fi
+
     git checkout preview --force
     git merge --no-ff master -m="preview < master [CI skip]"
     git push origin preview --force
